Create campaign
Created the system for generating a new campaign. The server will take the name of the campaign, and a password for later access, and save it all inside the campaigns folder. In that folder it will create a folder with the name given, and put a json file with the name and password in.