DM imageview
Maps uploaded to a campaign now get displayed on the address /dm/campaign. Images are loaded in with javascript, and put inside a div. Started working on a way for  resizing the images after they have been uploaded, but that is not completed. Going to work further on that later.