Invite link and code generation
Created a system for generating an ID and pin code for the players to use when logging in. This system still lacks a way of inputting the data, but that will be coming once the system for the DM is complete. That means I will have to make a way to display the data, and then make an interface for the players as well.