New mapstyle
Made the size of the maps displayed on the dm campaign page bigger. Also added seperate styling for phones, making the images even bigger. Also put the maps inside a div with the class of "Content-box". This makes a box around the map, and the new button I added called "Edit map". This button is not programmed yet, but that will come soon.