Check for invite link
Invite links generated by the DM will now be fetched and displayed on the Dm page if they are available. IDs and pin codes are generated by the server, and will be valid for 24 hours. After that they will be deleted next time someone generates new ones, or requests existing ones.