keep valid tokens
If a new login is initiated while a token is valid for the Campaign logins, the server will simpy resend the valid token. This allows for opening the page on several devices without everyone else being logged out. This is useful for me while testing, and there is probably someone out there that find it useful.

In any case, it makes the website no less secure, and might be useful in some cases. I see no reason for it not being implemented