Chat page for players
Started work on a chat page for the players. This chat will have the capability to roll dice, and run other commands in the future. For now it just adds what you send to the text box. Going to use a websocket interface to send data between all users, and cookies to distinguish people from each other on the serverside of things.